Add page with test filtered by status (pending, unplayed) Refactor Review
