+++ zucchini-ui-frontend/src/testRunFiltered/components/internal/TestRunFilteredTableContainer.js
@@ -1,12 +1,9 @@
 import { connect } from "react-redux";
-import { createSelector, createStructuredSelector } from "reselect";
+import { createStructuredSelector } from "reselect";
 
 import ScenarioTable from "../../../ui/components/ScenarioTable";
 
-const selectFiltered = createSelector(
-  state => state.testRunFiltered.testRunFiltered,
-  testRunFiltered => testRunFiltered
-);
+const selectFiltered = state => state.testRunFiltered.testRunFiltered;
 
 const selectProps = createStructuredSelector({
   scenarios: selectFiltered
